requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
drf-social-oauth2==2.2.0
filelock==3.13.4
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
platformdirs==4.2.1
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.0
